pyproject.toml
[tool.poetry]
name = "apielectionissues"
version = "0.1.0"
description = ""
authors = ["Jackson Severino da Rocha <jackson.severino.rocha@escola.pr.gov.br>"]
readme = "README.md"
packages = [
    {include = "api"},
    {include = "application"},
    {include = "domain"},
    {include = "infrastructure"},
    {include = "interfaces"},
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
sqlalchemy-utils = "^0.41.2"
alembic = "^1.13.2"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
python-multipart = "^0.0.9"
pyjwt = "^2.8.0"
python-jwt = "^4.1.0"
pylint = "^3.2.5"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
blue = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
